Fix typo in tests: use assert_called_once_with()
Fix a typo in tests whereas `.called_once_with()` is called rather than
`.assert_called_once_with()`.  This gets mocked in Python < 3.12
and therefore it is ignored (but doesn't test correctly) but it triggers
an error in Python 3.12:

```
E               AttributeError: 'called_once_with' is not a valid assertion. Use a spec for the mock if 'called_once_with' is meant to be an attribute.. Did you mean: 'assert_called_once_with'?
```